+++ starsky/starsky.foundation.sync/SyncServices/SyncSingleFile.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Globalization;
 using System.Linq;
 using System.Threading.Tasks;
 using starsky.foundation.database.Interfaces;
@@ -72,7 +73,7 @@ namespace starsky.foundation.sync.SyncServices
 				return await UpdateItem(dbItem, updatedDbItem.Size, subPath);
 			}
 
-			// to avoid resync
+			// to avoid reSync
 			updatedDbItem.Status = FileIndexItem.ExifStatus.OkAndSame;
 			AddDeleteStatus(statusItem, FileIndexItem.ExifStatus.DeletedAndSame);
 			
@@ -91,7 +92,7 @@ namespace starsky.foundation.sync.SyncServices
 			// route with database check
 			if ( _appSettings.ApplicationType == AppSettings.StarskyAppType.WebController )
 			{
-				_logger.LogInformation($"[SingleFile/db] {subPath}" );
+				_logger.LogInformation($"[SingleFile/db] {subPath} " + Synchronize.DateTimeDebug());
 			}
 
 			// Sidecar files are updated but ignored by the process
@@ -102,11 +103,11 @@ namespace starsky.foundation.sync.SyncServices
 
 			if ( statusItem.Status != FileIndexItem.ExifStatus.Ok )
 			{
-				_logger.LogInformation($"[SingleFile/db] status {statusItem.Status} for {subPath}");
+				_logger.LogInformation($"[SingleFile/db] status {statusItem.Status} for {subPath} {Synchronize.DateTimeDebug()}");
 				return statusItem;
 			}
 
-			var dbItem =  await _query.GetObjectByFilePathAsync(subPath);
+			var dbItem =  await _query.GetObjectByFilePathAsync(subPath, TimeSpan.FromSeconds(15));
 			// // // when item does not exist in Database
 			if ( dbItem == null )
 			{
@@ -120,12 +121,13 @@ namespace starsky.foundation.sync.SyncServices
 				return await UpdateItem(dbItem, updatedDbItem.Size, subPath);
 			}
 
-			// to avoid resync
+			// to avoid reSync
 			updatedDbItem.Status = FileIndexItem.ExifStatus.OkAndSame;
 			AddDeleteStatus(statusItem, FileIndexItem.ExifStatus.DeletedAndSame);
 			_logger.LogInformation($"[SingleFile/db] Same: {updatedDbItem.Status} for: {updatedDbItem.FilePath}");
 			return updatedDbItem;
 		}
+
 
 		/// <summary>
 		/// When the same stop checking and return value